Minor changes in reading config files and determining calib_dir
I only changed some minor things.

In Pose2Sim.py, the method for loading configuration files did not work when there was a file sorted before Config.toml. It now explicitly tries to load the Config.toml file.

In personAssociation.py and triangulation.py, the config_dir could not be determined correctly if there was a file sorted before the directory. It now only considers directories.